Adding runtime things to runtime package (#247)
* Move runtime types to proper package

* Cleanup

* Setting up MorphirRuntime and EvaluationLibrary to allow for PlatformSpecific behavior

* Reformat
