+++ morphir/runtime/jvm/src/org/finos/morphir/runtime/EvaluationLibraryPlatformSpecific.scala
@@ -1,5 +1,4 @@
-package org.finos.morphir
-package runtime
+package org.finos.morphir.runtime
 
 import org.finos.morphir.ir.Value.TypedValue
 import org.finos.morphir.ir.Value as V
@@ -15,16 +14,7 @@ import zio.json.*
 import org.finos.morphir.ir.json.MorphirJsonSupport.*
 import org.finos.morphir.runtime.quick.{EvaluatorQuick, Store}
 
-case class EvaluationLibrary(store: Store[Unit, Type.UType], modulePrefix: Option[String]) {
-  def runTest(moduleName: String, functionName: String, input: Any): Any = {
-    val fullName = modulePrefix match {
-      case Some(prefix) => s"$prefix:$moduleName:$functionName"
-      case None         => s"$moduleName:$functionName"
-    }
-    EvaluatorQuick.evalFunction(FQName.fromString(fullName), store, input)
-  }
-}
-object EvaluationLibrary {
+trait EvaluationLibraryPlatformSpecific {
   def apply(fileName: String, prefix: Option[String] = None): EvaluationLibrary = {
     val text = Source
       .fromFile(fileName)
@@ -38,5 +28,6 @@ object EvaluationLibrary {
     val store = Store.fromLibrary(library)
     EvaluationLibrary(store, prefix)
   }
+
   def apply(fileName: String, prefix: String): EvaluationLibrary = apply(fileName, Some(prefix))
 }
