+++ morphir/runtime/native/src/org/finos/morphir/runtime/EvaluationLibraryPlatformSpecific.scala
@@ -1,0 +1,3 @@
+package org.finos.morphir.runtime
+
+trait EvaluationLibraryPlatformSpecific {}
