+++ morphir/runtime/native/src/org/finos/morphir/runtime/MorphirRuntimePlatformSpecific.scala
@@ -1,0 +1,3 @@
+package org.finos.morphir.runtime
+
+trait MorphirRuntimePlatformSpecific {}
